+++ frontend/src/components/Homepage/index.js
@@ -3,6 +3,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import { NavLink } from 'react-router-dom';
 
 import { getBusinesses } from '../../store/business';
+
 // import { getOneBusinesses } from '../../store/business';
