+++ frontend/src/store/newBusiness.js
@@ -1,97 +1,102 @@
-import {csrfFetch} from './csrf'
+// import {csrfFetch} from './csrf'
 
 
-const ADD_BUINESS ='add business'
-const DELETE_BUINESS ='add business'
-const EDIT_BUINESS = 'edit business'
+// const ADD_BUINESS ='add business'
+// const DELETE_BUINESS ='add business'
+// const EDIT_BUINESS = 'edit business'
 
 
-const add = (business) =>({
-    type: ADD_BUINESS,
-    business,
-})
-const remove = (business) =>({
-    type: DELETE_BUINESS,
-    business,
-})
-const edit = (business) =>({
-    type: EDIT_BUINESS,
-    business,
-})
+// const add = (business) =>({
+//     type: ADD_BUINESS,
+//     business,
+// })
+// const remove = (business) =>({
+//     type: DELETE_BUINESS,
+//     business,
+// })
+// const edit = (business) =>({
+//     type: EDIT_BUINESS,
+//     business,
+// })
 
-export const addBusiness =({ownerId,title,description,address,city,state,zipCode}) => async (dispatch) =>{
-    const response = await csrfFetch(`/api/feature/new`,
-    {method:'POST',
-    headers:{'Content-Type' : 'application/json'},
-    body: JSON.stringify({ownerId,title,description,address,city,state,zipCode})
-    });
+// export const addBusiness =({ownerId,title,description,address,city,state,zipCode}) => async (dispatch) =>{
+//     const response = await csrfFetch(`/api/feature/new`,
+//     {method:'POST',
+//     headers:{'Content-Type' : 'application/json'},
+//     body: JSON.stringify({ownerId,title,description,address,city,state,zipCode})
+//     });
     
-    if (response.ok) {
-      const business = await response.json();
-      dispatch(add(business));
-      console.log(business)
-      return business
-    }
-}
+//     if (response.ok) {
+//       const business = await response.json();
+//       dispatch(add(business));
+//       console.log(business)
+//       return business
+//     }
+// }
 
-export const DeleteBusiness = (business) => async dispatch => {
-    console.log('>>>>>>>>>>>>>>>>>>|||||',business)
+// export const DeleteBusiness = (business) => async dispatch => {
+//     console.log('>>>>>>>>>>>>>>>>>>|||||',business)
   
-    const response = await csrfFetch(`/api/feature/${business.id}`,{
-      method:'DELETE',
-      headers:{'Content-Type' : 'application/json'},
-      body: JSON.stringify(business)
-    });
-    if (response.ok) {
-      await response.json();
-      dispatch(remove(business));
-      return business
-    }
-};
+//     const response = await csrfFetch(`/api/feature/${business.id}`,{
+//       method:'DELETE',
+//       headers:{'Content-Type' : 'application/json'},
+//       body: JSON.stringify(business)
+//     });
+//     if (response.ok) {
+//       await response.json();
+//       dispatch(remove(business));
+//       return business
+//     }
+// };
 
-export const EditBusiness = ({business}) => async (dispatch) =>{
+// export const EditBusiness = ({business}) => async (dispatch) =>{
   
-    console.log('>>>>>>>>>>>>>>>>>>',business)
+//     console.log('>>>>>>>>>>>>>>>>>>',business)
   
-    const response = await csrfFetch(`/api/fearure/${business.id}`,
-    {method:'PUT',
-    headers:{'Content-Type' : 'application/json'},
-    body: JSON.stringify(business)
-    });
+//     const response = await csrfFetch(`/api/feature/${business.id}`,
+//     {method:'PUT',
+//     headers:{'Content-Type' : 'application/json'},
+//     body: JSON.stringify(business)
+//     });
     
-    if (response.ok) {
-      await response.json();
-      dispatch(edit(business));
-      return business
-    }
-}
+//     if (response.ok) {
+//       await response.json();
+//       dispatch(edit(business));
+//       return business
+//     }
+// }
 
-const newBusinessReducer = (state = {}, action) => {
-    switch (action.type) {
-        case ADD_BUINESS: {
-            const newState = { ...state };
-            return newState;
-        }
-        case DELETE_BUINESS: {
-            const newState = { ...state };
-            const business = newState[action.buiness];
-            delete newState[business];
+// const newBusinessReducer = (state = {}, action) => {
+//     switch (action.type) {
+//         // case ADD_BUINESS: {
+//         //     const newState = { ...state };
+//         //     return newState;
+//         // }
+//         case ADD_BUINESS: {
+//             const newState = { ...state };
+//             newState[action.business.id] = action.business;
+//             return newState;
+//         }
+//         case DELETE_BUINESS: {
+//             const newState = { ...state };
+//             const business = newState[action.buiness];
+//             delete newState[business];
 
-            return newState;
-        }
-        case EDIT_BUINESS:{
-            const newState = { ...state };
-            const business = newState[action.buiness];
-            console.log('LLLLLL',business)
+//             return newState;
+//         }
+//         case EDIT_BUINESS:{
+//             const newState = { ...state };
+//             const business = newState[action.buiness];
+//             console.log('LLLLLL',business)
 
-            return newState
-          }
+//             return newState
+//           }
         
-      default:
-        return state;
-    }
-}
+//       default:
+//         return state;
+//     }
+// }
 
 
-export default newBusinessReducer
+// export default newBusinessReducer
